Generates UserAnalyzeImageCompleted events for tags where the image (with a different tag) has already been analyzed.
