+++ anchore_engine/services/catalog/catalog_impl.py
@@ -21,6 +21,7 @@ from anchore_engine.apis.exceptions import AnchoreApiError, BadRequest
 from anchore_engine.auth import aws_ecr
 from anchore_engine.clients import docker_registry
 from anchore_engine.clients.services import internal_client_for
+from anchore_engine.clients.services.catalog import CatalogClient
 from anchore_engine.clients.services.policy_engine import PolicyEngineClient
 from anchore_engine.db import (
     db_catalog_image,
@@ -34,6 +35,9 @@ from anchore_engine.db import (
 )
 from anchore_engine.services.catalog import utils
 from anchore_engine.subsys import logger, notifications, object_store, taskstate
+from anchore_engine.subsys.events.factories import (
+    analysis_complete_notification_factory,
+)
 from anchore_engine.util.docker import parse_dockerimage_string
 
 DeleteImageResponse = namedtuple("DeleteImageResponse", ["digest", "status", "detail"])
@@ -1964,7 +1968,9 @@ def add_or_update_image(
             dockerfile = None
             dockerfile_mode = None
 
-    # logger.debug("rationalized input for imageId ("+str(imageId)+"): " + json.dumps(image_ids, indent=4))
+    # Get a catalog client, for any events that need to be emitted
+    catalog_client = internal_client_for(CatalogClient, userId)
+
     addlist = {}
     for registry in list(image_ids.keys()):
         for repo in list(image_ids[registry].keys()):
@@ -2162,6 +2168,24 @@ def add_or_update_image(
                                 manifest = json.dumps({})
                             if not parent_manifest:
                                 parent_manifest = json.dumps({})
+
+                            # If image status is already analyzed, generate a UserAnalyzeImageCompleted event
+                            if image_record[
+                                "analysis_status"
+                            ] == anchore_engine.subsys.taskstate.complete_state(
+                                "analyze"
+                            ):
+                                for new_id in new_image_detail:
+                                    event = analysis_complete_notification_factory(
+                                        userId,
+                                        image_record["imageDigest"],
+                                        image_record["analysis_status"],
+                                        image_record["analysis_status"],
+                                        new_id,
+                                        annotations,
+                                    )
+                                    catalog_client.add_event(event)
+
                         except Exception as err:
                             raise anchore_engine.common.helpers.make_anchore_exception(
                                 err,
